+++ CHANGELOG.markdown
@@ -4,11 +4,16 @@ CHANGELOG for the Transifex API Package
 * 2.0.0
 
  * Dropped support for PHP 5.x, PHP 7.0 is the minimum requirement
+ * Replace `joomla/http` package with `guzzlehttp/guzzle` for HTTP adapter 
  * Add support for custom namespaces to `Transifex::get()`
  * Refactor API connector methods to return a full `Joomla\Http\Response` object
  * Removed magic getter in `Transifex` and associated class member vars
  * Support a default option in `Transifex::getOption()`
  * `Translationstrings::getStrings()` now typehints the `$options` parameter
+ * Rename `api.url` option to `base_uri`
+ * Remove support for `ArrayAccess` objects as an options param, must pass an array
+ * Removed `TransifexObject::fetchUrl()`, use the Guzzle API instead
+ * Removed `TransifexObject::processResponse()`, use the Guzzle API instead
 
 * 1.3.0 (2015-XX-XX)